teams, mini_league and game_week_players DB work. Including whole signup flow essentially in src/bin/test_teams_mini_leagues.rs except the discord part. Other work

+++ fpl_api/src/responses/mini_league.rs
@@ -15,33 +15,33 @@ pub struct League {
     pub name: String,
     pub created: DateTime<Utc>,
     pub closed: bool,
-    pub max_entries: Option<u32>,
+    pub max_entries: Option<i32>,
     pub league_type: String,
     pub scoring: String,
     pub admin_entry: TeamId,
     pub start_event: GameWeekId,
     pub code_privacy: String,
     pub has_cup: bool,
-    pub cup_league: Option<u32>,
-    pub rank: Option<u32>,
+    pub cup_league: Option<i32>,
+    pub rank: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct Standings {
     pub has_next: bool,
-    pub page: u32,
+    pub page: i32,
     pub results: Vec<Standing>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct Standing {
-    pub id: u32,
-    pub event_total: u8,
+    pub id: i32,
+    pub event_total: i16,
     pub player_name: String,
-    pub rank: u32,
-    pub last_rank: u32,
-    pub rank_sort: u32,
-    pub total: u16,
+    pub rank: i32,
+    pub last_rank: i32,
+    pub rank_sort: i32,
+    pub total: i16,
     pub entry: TeamId,
     pub entry_name: String,
     pub has_played: bool,
